chore: change local rep score type from u128 to u8

+++ core/handshake/src/dummy.rs
@@ -317,7 +317,7 @@ impl ReputationQueryInteface for MyReputationQuery {
     type SyncQuery = QueryRunner;
 
     /// Returns the reputation of the provided node locally.
-    fn get_reputation_of(&self, _peer: &NodePublicKey) -> Option<u128> {
+    fn get_reputation_of(&self, _peer: &NodePublicKey) -> Option<u8> {
         todo!()
     }
 }
